:art: refactor: Structure 변경
