+++ flutter/waitingnow/lib/src/app.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+import 'package:waitingnow/src/screen/home.dart';
+
+// stle 자동완성
+class MyApp extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return MaterialApp(
+      home: Home(),
+    );
+  }
+}
